requirements.txt
databricks-connect
sentence-transformers
faiss-cpu
requests
streamlit
